requirements.txt
# transformers==4.19.0
# annoy==1.17.0
# sentence-transformers==2.2.0
# llm>=0.1
azure-devops
crewai
duckduckgo-search
huggingface-hub
python-dotenv
langchain
pygithub
GitPython
loguru

pytest
pytest-coverage
pytest-dotenv
pytest-json-report


Django>=5.0
jsonfield
inflection
django-cors-headers
djangorestframework
python-dotenv
invoke
django-ninja
django-allauth
pytest-django
django-crispy-forms
crispy-bootstrap4
django_extensions
psycopg2-binary
pgvector
Pillow # Profile picture
celery
gunicorn
flower
scikit-learn
ruff
